Modify FriendVC and add FriendType enum file
